refactor: placeDetailDatasource 로직 이동
