+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/place/PlaceDataSourceImpl.kt
@@ -1,11 +1,14 @@
-package com.daedan.festabook.data.datasource.remote.placeDetail
+package com.daedan.festabook.data.datasource.remote.place
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.response.PlaceDetailResponse
+import com.daedan.festabook.data.model.response.PlaceResponse
 import com.daedan.festabook.data.service.PlaceService
 
-class PlaceDetailDataSourceImpl(
+class PlaceDataSourceImpl(
     private val placeService: PlaceService,
-) : PlaceDetailDataSource {
+) : PlaceDataSource {
+    override suspend fun fetchPlaces(): ApiResult<List<PlaceResponse>> = ApiResult.toApiResult { placeService.getPlaces() }
+
     override suspend fun fetchPlaceDetail(): ApiResult<PlaceDetailResponse> = ApiResult.toApiResult { placeService.getPlaceDetail() }
 }
