--- android/app/src/main/java/com/daedan/festabook/data/datasource/remote/placeDetail/PlaceDetailDataSource.kt
@@ -1,8 +1,0 @@
-package com.daedan.festabook.data.datasource.remote.placeDetail
-
-import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.PlaceDetailResponse
-
-interface PlaceDetailDataSource {
-    suspend fun fetchPlaceDetail(): ApiResult<PlaceDetailResponse>
-}
