--- android/app/src/main/java/com/daedan/festabook/data/datasource/remote/placeList/PlaceListDataSource.kt
@@ -1,8 +1,0 @@
-package com.daedan.festabook.data.datasource.remote.placeList
-
-import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.PlaceResponse
-
-interface PlaceListDataSource {
-    suspend fun fetchPlaces(): ApiResult<List<PlaceResponse>>
-}
