--- android/app/src/main/java/com/daedan/festabook/data/datasource/remote/placeList/PlaceListDataSourceImpl.kt
@@ -1,11 +1,0 @@
-package com.daedan.festabook.data.datasource.remote.placeList
-
-import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.PlaceResponse
-import com.daedan.festabook.data.service.PlaceService
-
-class PlaceListDataSourceImpl(
-    private val placeService: PlaceService,
-) : PlaceListDataSource {
-    override suspend fun fetchPlaces(): ApiResult<List<PlaceResponse>> = ApiResult.toApiResult { placeService.getPlaces() }
-}
