+++ android/app/src/main/java/com/daedan/festabook/data/repository/PlaceDetailRepositoryImpl.kt
@@ -1,16 +1,16 @@
 package com.daedan.festabook.data.repository
 
-import com.daedan.festabook.data.datasource.remote.placeDetail.PlaceDetailDataSource
+import com.daedan.festabook.data.datasource.remote.place.PlaceDataSource
 import com.daedan.festabook.data.model.response.toDomain
 import com.daedan.festabook.data.util.toResult
 import com.daedan.festabook.domain.model.PlaceDetail
 import com.daedan.festabook.domain.repository.PlaceDetailRepository
 
 class PlaceDetailRepositoryImpl(
-    private val placeDetailDataSource: PlaceDetailDataSource,
+    private val placeDataSource: PlaceDataSource,
 ) : PlaceDetailRepository {
     override suspend fun fetchPlaceDetail(): Result<PlaceDetail> {
-        val response = placeDetailDataSource.fetchPlaceDetail().toResult()
+        val response = placeDataSource.fetchPlaceDetail().toResult()
         return response.map { it.toDomain() }
     }
 }
